requirements.txt
flask
requests
flask-limiter
gunicorn
dnspython
beautifulsoup4